+++ src/VBox/Frontends/VirtualBox/src/wizards/newvd/UIWizardNewVDPageExpert.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIWizardNewVDPageExpert.cpp 44503 2013-02-01 06:28:53Z valery.portnyagin@oracle.com $ */
+/* $Id: UIWizardNewVDPageExpert.cpp 44511 2013-02-01 13:26:57Z noreply@oracle.com $ */
 /** @file
  *
  * VBox frontends: Qt4 GUI ("VirtualBox"):
@@ -191,7 +191,7 @@ void UIWizardNewVDPageExpert::sltMediumFormatChanged()
     ULONG uCapabilities = 0;
     QVector<KMediumFormatCapabilities> capabilities;
     capabilities = mf.GetCapabilities();
-    for (ULONG i = 0; i < capabilities.size(); i++)
+    for (int i = 0; i < capabilities.size(); i++)
         uCapabilities |= capabilities[i];
 
     bool fIsCreateDynamicPossible = uCapabilities & KMediumFormatCapabilities_CreateDynamic;
